requirements.txt
ollama>=0.3.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
pymupdf>=1.24.0
python-docx>=1.1.0
pyyaml>=6.0
rich>=13.7.0
prompt_toolkit>=3.0.43
pytest>=8.0.0
